backend/nlp/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
httpx==0.27.2
python-multipart==0.0.9
email-validator==2.2.0
spacy==3.8.2
transformers==4.45.2
torch==2.5.1
python-dateutil==2.9.0
langchain==0.3.0
langchain-community==0.3.0
openai==1.52.0
redis==5.1.1
celery==5.4.0
pydantic-settings==2.6.0
loguru==0.7.2